fix: move protocol to `base_url` env var to facilitate local dev
The pre-request script in the auth endpoint expected the protocol
to exist in the env var, yet all of the requests did not.
By moving the `https` protocol to the env var,
the user will not need to edit request URLs when running the requests
on localhost. Instead, they will only need to update the variable.
This also fixes the pre-request script.