[DPC-4758] remove references to deprecated IG (#151)
## 🎫 Ticket

https://jira.cms.gov/browse/DPC-4758

## 🛠 Changes

<!-- What was added, updated, or removed in this PR? -->
Removed links to outdated IG.

## ℹ️ Context

<!-- Why were these changes made? Add background context suitable for a
non-technical audience. -->

<!-- If any of the following security implications apply, this PR must
not be merged without Stephen Walter's approval. Explain in this section
and add @SJWalter11 as a reviewer.
  - Adds a new software dependency or dependencies.
  - Modifies or invalidates one or more of our security controls.
  - Stores or transmits data that was not stored or transmitted before.
- Requires additional review of security implications for other reasons.
-->

The BB2 IG was deprecated several years ago, yet is still referenced in
the DPC static site. Earlier in 2025, an audit was performed of static
site assets that led to the removal of the publication of the IG files
due to incompatibility with executive orders.


## 🧪 Validation

<!-- How were the changes verified? Did you fully test the acceptance
criteria in the ticket? Provide reproducible testing instructions and
screenshots if applicable. -->

---------

Co-authored-by: Christopher Ronning <127991809+chris-ronning-ny@users.noreply.github.com>
Co-authored-by: Christopher Ronning <cronning90@gmail.com>