DPC-4755 Migrate github actions to codebuild (#152)
## 🎫 Ticket

https://jira.cms.gov/browse/DPC-4766

## 🛠 Changes
* runs-on all use codebuild
* removed workflows for legacy AWS account
* removed static ig directory
* updated ci script to work under codebuild

## ℹ️ Context

Self-hosted on longer works in this repository, and we should not be
using github runners, so we need all the github actions for the static
repo to use codebuild.
I removed the legacy workflows to reduce the number of workflows that
needed to be updated.
I removed the ig directory because it was no longer necessary and was
generation warnings in the broken-link checker. (It looks like they were
added as part of the abandoned V2 move towards go:
https://jira.cms.gov/browse/DPC-1177).
Codebuild is a little slower and does not have the same API as ubuntu,
so some changes were necessary in the testing script.

## 🧪 Validation
broken-link-check, check_508_compliance, and ci-workflow all pass.